Add book BLOC
Add book bloc so all business logic is now done outside widgets.

see #5
