+++ pubspec.yaml
@@ -4,6 +4,7 @@ version: 1.0.0+1
 
 dependencies: 
   cupertino_icons: ^0.1.2
+  rxdart: ^0.22.1
   flutter: 
     sdk: flutter
